Implement LogDiff (with tests)
The LogDiff class handles emitting the Delete stream _minus_ any deletes
that have been marked successful.

This is a bit ugly because we've got worst case `NxM` performance here
where N is the delete log length and M is success log length.

This is preferable to pulling the entire log into memory?