Refactor the TestMemcachedConnector
The TestMemachedConnector now has a factory method.

This is the interface we will want for the "Real"
MemcachedConnector class. We want connect to be a factory
for connection objects.

Note the zaney use of static class variables to allow toggling
and testing around the behavior of the test connector.

We will want to use this behavior to validate the DeleteIssuer,
which is the "real" behavior that we want to test.