Quiet down warnings
There was an odd warning being shown from deep in the
memcached gem.

```
/usr/local/share/gems/gems/memcached-1.8.0/lib/memcached/exceptions.rb:53:
warning: instance variable @no_backtrace not initialized
```

That and a few other warnings about method overrides and circular
requires were creating a lot of noise in the output. Dropping the warn
level down a click suppresses the noise.